# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "用于家畜交易",
        "title": "淘牛马",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "phone_number 15294440097",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:13000",
    "paths": {
...
        "/chat/basic/create": {
            "post": {
                "description": "从现有用户基本信息中创建用户聊天账号信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户聊天基础"
                ],
                "summary": "创建用户聊天账号信息",
                "responses": {
                    "200": {
                        "description": "code\": 200, \"msg\": \"用户聊天账号信息创建成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\": 500, \"msg\": \"服务器内部错误\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/send_email_code": {
            "post": {
                "description": "Send an email verification code to the specified email address",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "公共方法"
                ],
                "summary": "发送邮箱验证码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email address to send the verification code to",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\":200,\"msg\":\"验证码已经成功发送\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "code\":400,\"msg\":\"请求参数错误\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\":500,\"msg\":\"服务器内部错误\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/send_phone_code": {
            "post": {
                "description": "向指定手机号发送验证码。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "公共方法"
                ],
                "summary": "发送手机验证码",
                "parameters": [
                    {
                        "description": "接收验证码的手机号",
                        "name": "phone",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/servers.Basic"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "code\":201,\"msg\":\"验证码已经成功发送\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "code\":400,\"msg\":\"请求无效。服务器无法理解请求\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\":500,\"msg\":\"服务器内部错误\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/user/chat/redisMsg": {
            "post": {
                "description": "从Redis中获取两个用户之间的聊天消息。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户私有方法"
                ],
                "summary": "从Redis获取用户聊天消息",
                "operationId": "getRedisMessages",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户A的ID",
                        "name": "userIdA",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "用户B的ID",
                        "name": "userIdB",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "消息起始位置",
                        "name": "start",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "消息结束位置",
                        "name": "end",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "是否反转消息顺序",
                        "name": "isRev",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/chat/sendMsg": {
            "get": {
                "description": "建立WebSocket连接并发送消息。",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "用户私有方法"
                ],
                "summary": "通过WebSocket发送消息",
                "operationId": "sendWebSocketMessage",
                "responses": {
                    "101": {
                        "description": "WebSocket连接已建立",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "WebSocket连接失败",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/chat/sendUserMsg": {
            "get": {
                "description": "通过WebSocket向指定用户发送消息。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户私有方法"
                ],
                "summary": "发送用户消息",
                "operationId": "sendUserMsg",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "接收者的用户ID",
                        "name": "userId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "servers.Basic": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
...
    }
}